+++ templates/index.php
@@ -1,49 +1,3 @@
-<?php
-$is_auth = rand(0,1);
-$user_name = 'Sergey'; 
-$user_avatar = 'img/user.jpg';
-$categories = ["Доски и лыжи", "Крепления", "Ботинки", "Одежда", "Инструменты", "Разное"]; 
-$lot_list = [
-    [ 'title' => '2014 Rossignol District Snowboard',
-      'category' => 'Доски и лыжи',
-      'price' => '10999,44',
-      'picture' => 'img/lot-1.jpg' ],
-
-    [ 'title' => 'DC Ply Mens 2016/2017 Snowboard',
-       'category' => 'Доски и лыжи',
-       'price' => '159999',
-       'picture' => 'img/lot-2.jpg' ],
-
-    [ 'title' => 'Крепления Union Contact Pro 2015 года размер L/XL',
-       'category' => 'Крепления',
-       'price' => '8000',
-       'picture' => 'img/lot-3.jpg' ],
-
-    [ 'title' => 'Ботинки для сноуборда DC Mutiny Charocal',
-       'category' => 'Ботинки',
-       'price' => '10999',
-       'picture' => 'img/lot-4.jpg' ],
-
-    [ 'title' => 'Куртка для сноуборда DC Mutiny Charocal',
-       'category' => 'Одежда',
-       'price' => '7500',
-       'picture' => 'img/lot-5.jpg' ],
-
-    [ 'title' => 'Маска Oakley Canopy',
-       'category' => 'Разное',
-       'price' => '5400',
-       'picture' => 'img/lot-6.jpg' ]];
-
-function priceform($price) 
-{
-    $number = ceil($price);
-    if ($number >= 1000) {
-        $number = number_format($number, 0, '.', ' ');
-    }
-    return $number.' <b class="rub">р</b>';
-}
-?>
-<main class="container">
 <section class="promo">
     <h2 class="promo__title">Нужен стафф для катки?</h2>
     <p class="promo__text">На нашем интернет-аукционе ты найдёшь самое эксклюзивное сноубордическое и горнолыжное снаряжение.</p>
@@ -82,4 +36,3 @@ function priceform($price)
         <?php endforeach; ?>
     </ul>
 </section>
-</main>
